+++ src/Modules/MayaPro.WarehouseApi.Modules.Expenses/Application/UseCases/CreateExpenseType/CreateExpenseTypeValidator.cs
@@ -1,0 +1,12 @@
+using FluentValidation;
+
+namespace MayaPro.WarehouseApi.Modules.Expenses.Application.UseCases.CreateExpenseType;
+
+public sealed class CreateExpenseTypeValidator : AbstractValidator<CreateExpenseTypeCommand>
+{
+    public CreateExpenseTypeValidator()
+    {
+        RuleFor(x => x.Name)
+            .NotEmpty().WithMessage("Xərc növü adı boş ola bilməz");
+    }
+}